Remove almost everything
We are starting more-or-less over from scratch as we are
developing an new way of constructing the state spaces. Since so
much depends on the state spaces, this will break almost all of
the tests. It is best to just remove the tests and start again
fresh as we reimplement the various features.