Fix monotonicity of relative entropy
When we perform swaps at the macroscale, we have to realize that the
swaps are actually happening at the microscale. If we swap the macrostate
probabilities, then we end up with non-monotonic relative entropy! To
resolve the issue, the probabilities of the microstates in a given
macrostate are assumed to be uniformly distributed. When a swap occurs,
instead of swapping the macrostate probabilites, we swap these "max-ent"
style microstate probabilities.

This recovers the monotonicity property.