Fix a major performance bug
When we were looking for swaps, we would go down the list of state
probabilities and subtract them from the randomly selected probability
until would be come negative. This is was a very costly process because it
required creating a new BigFloat with each loop. As a fix, we construct an
array of the cummulative sums of the probabilities as Context.Pcum. Then
we simply search for the first one cummulative probability that is a
least equal to it.

This results is a HUGE performance improvement. In the case of simulating
the microscale dynamics of the minimal, 4-monomer system, the run time
reduced from 239min to 7min, i.e. a factor of better than 30!